+++ 02_daemonset.tf
@@ -369,8 +369,8 @@ resource "kubernetes_daemon_set_v1" "daemonset" {
             }
 
             resources {
-              requests = init_container.value.resources[local.infrastructureSize].requests
-              limits   = init_container.value.resources[local.infrastructureSize].limits
+              requests = lookup(var.infraOverrideConfig.resources, init_container.key, lookup(init_container.value.resources, local.infrastructureSize, local.fallbackResources)).requests
+              limits   = lookup(var.infraOverrideConfig.resources, init_container.key, lookup(init_container.value.resources, local.infrastructureSize, local.fallbackResources)).limits
             }
 
             dynamic "volume_mount" {
@@ -763,8 +763,8 @@ resource "kubernetes_daemon_set_v1" "daemonset" {
             }
 
             resources {
-              requests = { for k, v in container.value.resources[local.infrastructureSize].requests : k => v == null ? null : "${regex(local.resourceMultiplierRegex, v)[0] * local.resourceMultiplier}${regex(local.resourceMultiplierRegex, v)[1]}" }
-              limits   = { for k, v in container.value.resources[local.infrastructureSize].limits : k => v == null ? null : "${regex(local.resourceMultiplierRegex, v)[0] * local.resourceMultiplier}${regex(local.resourceMultiplierRegex, v)[1]}" }
+              requests = lookup(var.infraOverrideConfig.resources, container.key, lookup(container.value.resources, local.infrastructureSize, local.fallbackResources)).requests
+              limits   = lookup(var.infraOverrideConfig.resources, container.key, lookup(container.value.resources, local.infrastructureSize, local.fallbackResources)).limits
             }
 
             dynamic "volume_mount" {