+++ 03_statefulset.tf
@@ -11,7 +11,7 @@ resource "kubernetes_stateful_set_v1" "statefulset" {
 
   spec {
     service_name          = kubernetes_service_v1.clusterIp.0.metadata.0.name
-    replicas              = var.podResourceTypeConfig.minReplicas
+    replicas = var.infraOverrideConfig.replicas.min == null ? var.podResourceTypeConfig.minReplicas : var.infraOverrideConfig.replicas.min
     pod_management_policy = var.podResourceTypeConfig.podManagementPolicy
 
     selector {
@@ -388,8 +388,8 @@ resource "kubernetes_stateful_set_v1" "statefulset" {
             }
 
             resources {
-              requests = init_container.value.resources[local.infrastructureSize].requests
-              limits   = init_container.value.resources[local.infrastructureSize].limits
+              requests = lookup(var.infraOverrideConfig.resources, init_container.key, lookup(init_container.value.resources, local.infrastructureSize, local.fallbackResources)).requests
+              limits   = lookup(var.infraOverrideConfig.resources, init_container.key, lookup(init_container.value.resources, local.infrastructureSize, local.fallbackResources)).limits
             }
 
             dynamic "volume_mount" {
@@ -790,8 +790,8 @@ resource "kubernetes_stateful_set_v1" "statefulset" {
             }
 
             resources {
-              requests = { for k, v in container.value.resources[local.infrastructureSize].requests : k => v == null ? null : "${regex(local.resourceMultiplierRegex, v)[0] * local.resourceMultiplier}${regex(local.resourceMultiplierRegex, v)[1]}" }
-              limits   = { for k, v in container.value.resources[local.infrastructureSize].limits : k => v == null ? null : "${regex(local.resourceMultiplierRegex, v)[0] * local.resourceMultiplier}${regex(local.resourceMultiplierRegex, v)[1]}" }
+              requests = lookup(var.infraOverrideConfig.resources, container.key, lookup(container.value.resources, local.infrastructureSize, local.fallbackResources)).requests
+              limits   = lookup(var.infraOverrideConfig.resources, container.key, lookup(container.value.resources, local.infrastructureSize, local.fallbackResources)).limits
             }
 
             dynamic "volume_mount" {