+++ 19_variables.tf
@@ -11,6 +11,26 @@ variable "architecture" {}
 
 variable "infrastructureSize" {}
 
+variable "infraOverrideConfig" {
+  type = object({
+    replicas = optional(object({
+      min = optional(number, null)
+      max = optional(number, null)
+    }), {})
+    resources = optional(map(object({
+      requests = optional(object({
+        cpu = optional(string, null)
+        memory = optional(string, null)
+      }), {})
+      limits = optional(object({
+        cpu    = optional(string, null)
+        memory = optional(string, null)
+      }), {})
+    })), {})
+  })
+  default = {}
+}
+
 variable "consistency" {
   type = object({
     hard = object({